+++ src/e2e/app/cli/sqlite/tests/tests/mod.rs
@@ -10,4 +10,5 @@
 mod test_auth;
 mod test_rest_api;
 mod test_selftest;
+mod test_smtp;
 mod test_system_generate_token;